core:
- синхронизировал коммит r96 с тест-зоной
- поправил тип данных ( uint32 )

git-svn-id: svn://300murlocs.com/core/stable/code@112 b8eaa9ba-51a4-de11-8768-001a646d8106

+++ src/game/GameObject.cpp
@@ -870,7 +870,7 @@ void GameObject::Use(Unit* user)
                 if( (gofaction->IsHostileTo(*plrfaction) ) )
                 {
                     std::stringstream gobid;
-                    gobid << "Faction hack (GO entry: " << mygoinfo->id << ", GO faction: " << gofact << "; Player race: " << plr->getRace() << ")";
+                    gobid << "Faction hack (GO entry: " << mygoinfo->id << ", GO faction: " << gofact << "; Player race: " << uint32(plr->getRace()) << ")";
 
                     std::stringstream Position;
                     Position << "Player Position: " << plr->GetPositionX() << " " << plr->GetPositionY() << " "
